+++ sphinx_gitref/__main__.py
@@ -1,0 +1,4 @@
+from .commands import invoke
+
+if __name__ == "__main__":
+    invoke()